updated signing to new structure, and added utils-cli from mix pr
