+++ src/cli.ts
@@ -3,104 +3,20 @@
 /* NOTE: calling this file entropy.ts helps commander parse process.argv */
 import { Command, Option } from 'commander'
 import launchTui from './tui'
-import * as config from './config'
 import { EntropyTuiOptions } from './types'
 
 import { cliListAccounts } from './flows/manage-accounts/cli'
-import { getSelectedAccount, stringify } from './common/utils'
 import Entropy from '@entropyxyz/sdk'
-import { initializeEntropy } from './common/initializeEntropy'
 import { BalanceCommand } from './balance/command'
 import { TransferCommand } from './transfer/command'
-import { SigningCommand } from './signing/command'
+import { entropySignCommand } from './signing/command'
+import { currentAccountAddressOption, endpointOption, loadEntropy, passwordOption, cliWrite } from './common/utils-cli'
 
 const program = new Command()
 // Array of restructured commands to make it easier to migrate them to the new "flow"
 const RESTRUCTURED_COMMANDS = ['balance']
 
-function endpointOption (){
-  return new Option(
-    '-e, --endpoint <endpoint>',
-    [
-      'Runs entropy with the given endpoint and ignores network endpoints in config.',
-      'Can also be given a stored endpoint name from config eg: `entropy --endpoint test-net`.'
-    ].join(' ')
-  )
-    .env('ENDPOINT')
-    .argParser(aliasOrEndpoint => {
-      /* see if it's a raw endpoint */
-      if (aliasOrEndpoint.match(/^wss?:\/\//)) return aliasOrEndpoint
-
-      /* look up endpoint-alias */
-      const storedConfig = config.getSync()
-      const endpoint = storedConfig.endpoints[aliasOrEndpoint]
-      if (!endpoint) throw Error('unknown endpoint alias: ' + aliasOrEndpoint)
-
-      return endpoint
-    })
-    .default('ws://testnet.entropy.xyz:9944/')
-    // NOTE: argParser is only run IF an option is provided, so this cannot be 'test-net'
-}
-
-function passwordOption (description?: string) {
-  return new Option(
-    '-p, --password <password>',
-    description || 'Password for the account'
-  )
-}
-
-function currentAccountAddressOption () {
-  const storedConfig = config.getSync()
-  return new Option(
-    '-a, --account <accountAddress>',
-    'Sets the current account for the session or defaults to the account stored in the config'
-  )
-    .env('ACCOUNT_ADDRESS')
-    .argParser(async (address) => {
-      if (address === storedConfig.selectedAccount) return address
-      // Updated selected account in config with new address from this option
-      const newConfigUpdates = { selectedAccount: address }
-      await config.set({ ...storedConfig, ...newConfigUpdates })
-
-      return address
-    })
-    .hideHelp()
-    .default(storedConfig.selectedAccount)
-}
-
 let entropy: Entropy
-
-export async function loadEntropy (address: string, endpoint: string, password?: string): Promise<Entropy> {
-  const storedConfig = config.getSync()
-  const selectedAccount = getSelectedAccount(storedConfig.accounts, address)
-
-  if (!selectedAccount) throw new Error(`AddressError: No account with address ${address}`)
-
-  // check if data is encrypted + we have a password
-  if (typeof selectedAccount.data === 'string' && !password) {
-    throw new Error('AuthError: This account requires a password, add --password <password>')
-  }
-
-  entropy = await initializeEntropy({ keyMaterial: selectedAccount.data, endpoint, password })
-
-  if (!entropy?.keyring?.accounts?.registration?.pair) {
-    throw new Error("Signer keypair is undefined or not properly initialized.")
-  }
-
-  return entropy
-}
-
-async function reloadEntropy (newAddress: string, oldAddress: string, endpoint: string): Promise<void> {
-  try {
-    entropy = await loadEntropy(newAddress, endpoint)
-  } catch (error) {
-    if (error.message.includes('AddressError')) {
-      entropy = await loadEntropy(oldAddress, endpoint)
-      return
-    }
-    throw error
-  }
-}
 
 /* no command */
 program
@@ -121,10 +37,10 @@ program
       // balance includes an address argument, use that address to instantiate entropy
       // can keep the conditional to check for length of args, and use the first index since it is our pattern to have the address as the first argument
       if (RESTRUCTURED_COMMANDS.includes(actionCommand.name()) && actionCommand.args.length) {
-        await loadEntropy(actionCommand.args[0], actionCommand.opts().endpoint, actionCommand.opts().password)
+        entropy = await loadEntropy(entropy, actionCommand.args[0], actionCommand.opts().endpoint, actionCommand.opts().password)
       } else {
         // if address is not an argument, use the address from the option
-        await loadEntropy(actionCommand.opts().account, actionCommand.opts().endpoint, actionCommand.opts().password)
+        entropy = await loadEntropy(entropy, actionCommand.opts().account, actionCommand.opts().endpoint, actionCommand.opts().password)
       }
     }
   })
@@ -139,7 +55,7 @@ program.command('list')
   .action(async () => {
     // TODO: test if it's an encrypted account, if no password provided, throw because later on there's no protection from a prompt coming up
     const accounts = await cliListAccounts()
-    writeOut(accounts)
+    cliWrite(accounts)
     process.exit(0)
   })
 
@@ -152,7 +68,7 @@ program.command('balance')
   .action(async (address, opts) => {
     const balanceCommand = new BalanceCommand(entropy, opts.endpoint)
     const balance = await balanceCommand.getBalance(address)
-    writeOut(balance)
+    cliWrite(balance)
     process.exit(0)
   })
 
@@ -168,47 +84,11 @@ program.command('transfer')
   .action(async (_source, destination, amount, opts) => {
     const transferCommand = new TransferCommand(entropy, opts.endpoint)
     await transferCommand.sendTransfer(destination, amount)
-    // writeOut(??) // TODO: write the output
+    // cliWrite(??) // TODO: write the output
     process.exit(0)
   })
 
 /* Sign */
-program.command('sign')
-  .description('Sign a message using the Entropy network. Output is a signature (string)')
-  .argument('[address]', 'Entropy Account address to be used to sign')
-  .argument('[message]', 'Message you would like to sign')
-  .argument('[msgPath]', 'Path to file of content you would like to sign')
-  .addOption(passwordOption('Password for the source account (if required)'))
-  .addOption(endpointOption())
-  .addOption(currentAccountAddressOption())
-  .addOption(
-    new Option(
-      '-r, --raw',
-      'Signs the provided message using the Raw Signing method. Output is a signature (string)'
-    )
-  )
-  .action(async (address, message, msgPath, opts) => {
-    if (address) {
-      await reloadEntropy(
-        address,
-        entropy.keyring.accounts.registration.address, opts.endpoint
-      )
-    }
-    const signingCommand = new SigningCommand(entropy, opts.endpoint)
-    // TO-DO: Add ability for raw signing here, maybe? new raw option can be used for the conditional
-    /**
-     * if (opts.raw) {
-     *   implement raw sign here
-     * }
-     */
-    const signature = await signingCommand.signMessage({ msg: message, msgPath })
-    writeOut(signature)
-    process.exit(0)
-  })
-
-function writeOut (result) {
-  const prettyResult = stringify(result)
-  process.stdout.write(prettyResult)
-}
+entropySignCommand(entropy, program)
 
 program.parseAsync().then(() => {})
